Refactor: Update orchestration and Docker configurations for .NET 10 support; remove obsolete files and enhance logging
