+++ test/SmokeTests/BackendSmokeTests/Netherite/Dockerfile
@@ -12,5 +12,6 @@ RUN cd /root/test/SmokeTests/BackendSmokeTests/Netherite && \
 # Deploy the app
 FROM mcr.microsoft.com/azure-functions/dotnet-isolated:4-dotnet-isolated10.0
 ENV AzureWebJobsScriptRoot=/home/site/wwwroot \
-    AzureFunctionsJobHost__Logging__Console__IsEnabled=true
+    AzureFunctionsJobHost__Logging__Console__IsEnabled=true \
+    FUNCTIONS_WORKER_RUNTIME=dotnet-isolated
 COPY --from=build-env ["/home/site/wwwroot", "/home/site/wwwroot"]
