+++ test/SmokeTests/BackendSmokeTests/Netherite/Program.cs
@@ -1,0 +1,7 @@
+using Microsoft.Extensions.Hosting;
+
+var host = new HostBuilder()
+    .ConfigureFunctionsWorkerDefaults()
+    .Build();
+
+host.Run();
