+++ test/e2e/Tests/Tests/DistributedTracingEntitiesTests.cs
@@ -14,6 +14,7 @@ public class DistributedTracingEntitiesTests
 {
     private readonly FunctionAppFixture _fixture;
     private readonly ActivityListener _activityListener;
+    private static readonly ActivitySource activitySource = new ActivitySource("DistributedTracingEntitiesTests");
 
     public DistributedTracingEntitiesTests(FunctionAppFixture fixture, ITestOutputHelper testOutputHelper)
     {
@@ -45,7 +46,6 @@ public class DistributedTracingEntitiesTests
     public async Task DistributedTracingEntitiesTest()
     {
         // Start an Activity here which will appear as the root of all other Activities generated by this test
-        ActivitySource activitySource = new ActivitySource("DistributedTracingEntitiesTests");
         using Activity? activity = activitySource.StartActivity("OrchestrationTests");
 
         Assert.NotNull(activity);
@@ -89,7 +89,6 @@ public class DistributedTracingEntitiesTests
     public async Task DistributedTracingEntitiesFromClientTest()
     {
         // Start an Activity here which will appear as the root of all other Activities generated by this test
-        ActivitySource activitySource = new ActivitySource("DistributedTracingEntitiesTests");
         using Activity? activity = activitySource.StartActivity("ClientTests");
 
         Assert.NotNull(activity);
